+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Creating pack not resetting modloader when switched back to none
+- Issue with file downloads for non third party enabled files on CurseForge not working
 
 ### Misc
 - Remove old FTB pack installing code